BEAM-6627 Move IOITMetrics to a common test package, move BQ options to common IOIT option class
